Moved `from_layer` logic to AutoModel
